Cleanup code for pull request (style & remove unused full gradle configuration from New project)
