+++ build.gradle
@@ -18,7 +18,6 @@ repositories {
 }
 
 dependencies {
-    compile 'org.apache.commons:commons-text:1.9'
     testImplementation 'org.junit.jupiter:junit-jupiter-api:5.6.0'
     testRuntimeOnly 'org.junit.jupiter:junit-jupiter-engine'
 }
