+++ src/main/resources/fileTemplates/internal/Frege Module.fr.ft
